backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.7.1
cron-descriptor==1.4.5
debugpy==1.8.13
Django==5.1.7
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.15.2
docker==7.1.0
factory_boy==3.3.3
Faker==37.1.0
ddtrace==2.8.7
idna==3.10
iniconfig==2.1.0
kombu==5.5.0
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2==2.9.10
pytest==8.3.5
pytest-django==4.10.0
pytest-mock==3.14.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-mock==1.12.1
six==1.17.0
sqlparse==0.5.3
tenacity==9.0.0
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
